Type : Fixes Desc : Last Min Fixes. Fixed Readme. Generated Video and source code pdf.
